renamed game_setup to lobby_admin
