+++ app/locales/fr/translation.json
@@ -72,8 +72,8 @@
 		"password": "Mot de passe",
 		"bio": "Bio"
 	},
-	"game_setup": {
-		"game_setup": "Configuration du jeu",
+	"lobby_admin": {
+		"lobby_admin": "Configuration du jeu",
 		"general_settings": "Paramètres généraux",
 		"map": "Carte",
 		"modes": {
