+++ app/pages/layouts/menu/loggedin.ejs
@@ -7,7 +7,7 @@
 	>
 	Chat
 	</button>
-	<button data-page="game_setup" class="cursor-pointer hover:scale-110 hover:glow-blue text-white font-bold py-2 px-4 rounded-full glow-green"
+	<button data-page="lobby_admin" class="cursor-pointer hover:scale-110 hover:glow-blue text-white font-bold py-2 px-4 rounded-full glow-green"
 		onclick="createGame()"><%= t('menu.game') %></button>
 	<button data-page="logout" class="cursor-pointer hover:scale-110 hover:glow-blue text-white font-bold py-2 px-4 rounded-full glow-green"
 		onclick="logout()">
