+++ app/src/routes/partial/pages/index.ts
@@ -151,7 +151,7 @@ const pages: FastifyPluginAsync = async (fastify, opts): Promise<void> => {
 						profile.id,
 						fastify
 					);
-				} else if (page === 'game_setup') {
+				} else if (page === 'lobby_admin') {
 					const user = await checkAuth(req, true, fastify);
 					if (!user)
 						return reply.code(401).send({ error: 'Unauthorized' });
